% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionariocv.R
\name{diccionario_cv}
\alias{diccionario_cv}
\title{Diccionario de datos cobertura vegetal}
\usage{
diccionario_cv(filename)
}
\arguments{
\item{filename}{sf path}
}
\value{
objecto de tipo sf
}
\description{
Diccionario de datos cobertura vegetal
}
