% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionariodelim.R
\name{diccionario_delim}
\alias{diccionario_delim}
\title{Diccionario de datos Propuesta de establecimiento de la ZPCE}
\usage{
diccionario_delim(filename)
}
\arguments{
\item{filename}{file path}
}
\value{
sf
}
\description{
Diccionario de datos Propuesta de establecimiento de la ZPCE
}
