% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionarioean.R
\name{diccionario_ean}
\alias{diccionario_ean}
\title{Diccionario espacios con afectaciones normativas PDU 2013 - 2023}
\usage{
diccionario_ean(filename)
}
\arguments{
\item{filename}{Shapefile}
}
\value{
sf
}
\description{
Diccionario espacios con afectaciones normativas PDU 2013 - 2023
}
