% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionarioeen.R
\name{diccionario_een}
\alias{diccionario_een}
\title{Diccionario de datos ecosistemas y espacios naturales}
\usage{
diccionario_een(filename)
}
\arguments{
\item{filename}{path to shp file}
}
\value{
sf
}
\description{
Diccionario de datos ecosistemas y espacios naturales
}
