% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionarioga.R
\name{diccionario_ga}
\alias{diccionario_ga}
\title{Diccionario de datos grado de antropizacion}
\usage{
diccionario_ga(filename)
}
\arguments{
\item{filename}{path to shp file}
}
\value{
sf
}
\description{
Diccionario de datos grado de antropizacion
}
