% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacioncv.R
\name{review_cv_valid}
\alias{review_cv_valid}
\title{Validación de las variables tipo de cobertura y cobertura vegetal}
\usage{
review_cv_valid(sf)
}
\arguments{
\item{sf}{Objeto sf}
}
\value{
text string
}
\description{
Validación de las variables tipo de cobertura y cobertura vegetal
}
