% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp_write.R
\name{shape_write}
\alias{shape_write}
\title{Exportar shapefiles}
\usage{
shape_write(sf, folder, filename)
}
\arguments{
\item{sf}{objeto sf}

\item{folder}{path dela carpeta donde se guardara el archivo}

\item{filename}{nombre con el cual se guardara el archivo}
}
\value{
Save shapefile
}
\description{
Exportar shapefiles
}
