% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{str_extract_after}
\alias{str_extract_after}
\title{Extract part of string}
\usage{
str_extract_after(string, pattern, which = "first", num_char = NULL)
}
\arguments{
\item{string}{string}

\item{pattern}{pattern to find}

\item{which}{position}

\item{num_char}{number of characters}
}
\description{
Extract part of string
}
\keyword{internal}
