% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacioncv.R
\name{validacion_cv}
\alias{validacion_cv}
\title{Validación de las variables tipo de cobertura y cobertura vegetal}
\usage{
validacion_cv(sf)
}
\arguments{
\item{sf}{Objeto resultante de la función 'diccionari_cv'}
}
\value{
Message
}
\description{
Validación de las variables tipo de cobertura y cobertura vegetal
}
\examples{
# diccionario_cv(filename) |>
# validacion_cv()
validacion de tipo de cobertura
}
