% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionariocv.R
\name{diccionario_cv}
\alias{diccionario_cv}
\title{Diccionario cobertura vegetal}
\usage{
diccionario_cv(filename)
}
\arguments{
\item{filename}{Path del archivo .shp}
}
\value{
sf object
}
\description{
Diccionario cobertura vegetal
}
\examples{
seleccionar el nombre de la columna que contiene la informacion
del tipo de cobertura NATURAL - ANTROPICA
seleccionar el nombre de la columna que contiene la informacion
de los tipos de cobertura vegetal
ARBOREA - HERBAZAL - MATORRAL- PASTIZAL...
seleccione de columnas presentes en el shapefile
recuperar columnas presentes en el shp
renombrar las columnas
validar los niveles de cobertura vegetal
}
