% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionariopropzoni.R
\name{diccionario_propzoni}
\alias{diccionario_propzoni}
\title{Diccionario propuesta de zonificacion}
\usage{
diccionario_propzoni(filename)
}
\arguments{
\item{filename}{File path}
}
\value{
sf
}
\description{
Diccionario de los shapefiles de las propuestas de zonificacion:
\itemize{
\item \strong{ZPCE} ZONAS DE PROTECCION Y CONSERVACION ECOLOGICA
\item \strong{ZIERE} ZONAS DE INTERVENCION ESPECIESL Y RECUPERACION DE ECOSISTEMAS
\item \strong{ZPRH} ZONAS DE PROTECCION DEL RECURSO HIDRICO
}
}
